pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-mvp-factory"
version = "0.1.0"
description = "CLI-first council runner for multi-model debate with HITL approval"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
    "langgraph>=0.2.0",
    "pyyaml>=6.0.0",
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
council = "agentic_mvp_factory.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

